+++ src/components/atoms/LargeGrid.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react'
 
-export default function CommunityGrid({ children }: PropsWithChildren<{}>) {
+export default function LargeGrid({ children }: PropsWithChildren<{}>) {
   return (
     <div className="grid grid-cols-1 md:grid-cols-3 gap-4 m-4">
       {children}
