+++ src/components/organisms/CommunitiesList.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 
 import Button from '@src/components/atoms/Button'
-import CommunityGrid from '@src/components/molecules/CommunityGrid'
+import LargeGrid from '@src/components/atoms/LargeGrid'
 import CommunityTile from '@src/components/molecules/CommunityTile'
 import { CommunityDocument } from '@src/model/Community'
 
@@ -15,7 +15,7 @@ export default function CommunitiesList({ communities }: CommunitiesListProps) {
       <div className="h-9 flex justify-end mt-4">
         <Button href="/communities/create">Create new...</Button>
       </div>
-      <CommunityGrid>
+      <LargeGrid>
         {communities.map(doc => (
           <Link key={doc.id} href={`/communities/${doc.id}`}>
             <a>
@@ -23,7 +23,7 @@ export default function CommunitiesList({ communities }: CommunitiesListProps) {
             </a>
           </Link>
         ))}
-      </CommunityGrid>
+      </LargeGrid>
     </>
   )
 }
